+++ app/components/forms/AppFormField.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useFormikContext } from 'formik'
 
-import AppTextInput from "../components/AppTextInput"
-import ErrorMessage from '../components/ErrorMessage';
+import AppTextInput from "../AppTextInput"
+import ErrorMessage from './ErrorMessage';
 
 function AppFormField({ name, ...otherProps }) {
