+++ app/screens/LoginScreen.js
@@ -1,17 +1,9 @@
 import React, { useState } from 'react';
 import { StyleSheet, Image } from 'react-native';
-import { Formik } from 'formik'
 import * as Yup from 'yup'
 
 import Screen from "../components/screen"
-import AppTextInput from "../components/AppTextInput"
-import AppButton from '../components/AppButton';
-import AppText from "../components/AppText"
-import ErrorMessage from '../components/ErrorMessage';
-import AppFormField from '../components/AppFormField';
-import SubmitButton from '../components/SubmitButton';
-import AppForm from '../components/AppForm';
-
+import { AppForm, AppFormField, SubmitButton } from '../components/forms'
 
 const validationSchema = Yup.object().shape({
     email: Yup.string().required().email().label("Email"),
